Extract CHA from booster-task-analyser as a standalone module
