+++ booster-cha/README.md
@@ -1,0 +1,3 @@
+# booster-cha
+
+This module is used for CHA (class hierarchy analysis)
