+++ booster-cha/build.gradle
@@ -1,0 +1,5 @@
+apply from: '../gradle/booster.gradle'
+
+dependencies {
+    implementation project(':booster-transform-asm')
+}
