+++ booster-cha/src/main/kotlin/com/didiglobal/booster/cha/AbstractClassSet.kt
@@ -1,4 +1,4 @@
-package com.didiglobal.booster.task.analyser.cha
+package com.didiglobal.booster.cha
 
 import com.didiglobal.booster.transform.asm.className
 import org.objectweb.asm.tree.ClassNode
