+++ booster-cha/src/main/kotlin/com/didiglobal/booster/cha/ClassHierarchy.kt
@@ -1,4 +1,4 @@
-package com.didiglobal.booster.task.analyser.cha
+package com.didiglobal.booster.cha
 
 import com.didiglobal.booster.transform.asm.isFinal
 import com.didiglobal.booster.transform.asm.isInterface
