+++ booster-cha/src/test/kotlin/com/didiglobal/booster/cha/graph/dot/WebSafeColorPaletteTest.kt
@@ -1,5 +1,6 @@
-package com.didiglobal.booster.task.analyser.palette
+package com.didiglobal.booster.cha.graph.dot
 
+import com.didiglobal.booster.cha.graph.dot.WebSafeColorPalette
 import kotlin.test.Test
 import kotlin.test.assertNotEquals
