+++ booster-task-analyser/build.gradle
@@ -4,6 +4,7 @@ dependencies {
     kapt "com.google.auto.service:auto-service:1.0-rc6"
     implementation project(':booster-aapt2')
     implementation project(':booster-api')
+    implementation project(':booster-cha')
     implementation project(':booster-transform-asm')
     compileOnly 'com.android.tools.build:gradle:3.0.0'
     testCompileOnly 'com.android.tools.build:gradle:3.0.0'
