+++ settings.gradle
@@ -19,6 +19,7 @@ include ':booster-android-gradle-v3_6'
 include ':booster-annotations'
 include ':booster-api'
 include ':booster-build'
+include ':booster-cha'
 include ':booster-command'
 include ':booster-gradle-plugin'
 include ':booster-task-analyser'
